+++ src/hooks/LevelSearchLayer.cpp
@@ -13,7 +13,7 @@ class $modify(CbHookLevelSearchLayer, LevelSearchLayer) {
         if (auto search = getChildByID("search-bar")) {
             log::trace("node {} found", search->getID());
 
-            if (auto menu = typeinfo_cast<ClipboardMenu*>(search->getChildByID("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.75f);
+            if (auto menu = typeinfo_cast<ClipboardMenu*>(search->getChildByID("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.5f);
         };
 
         return true;