+++ src/hooks/SecretLayer.cpp
@@ -13,7 +13,7 @@ class $modify(CbHookSecretLayer, SecretLayer) {
     bool init() {
         if (!SecretLayer::init()) return false;
 
-        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.875f);
+        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.75f);
 
         return true;
     };
@@ -23,7 +23,7 @@ class $modify(CbHookSecretLayer2, SecretLayer2) {
     bool init() {
         if (!SecretLayer2::init()) return false;
 
-        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.875f);
+        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.75f);
 
         return true;
     };
@@ -33,7 +33,7 @@ class $modify(CbHookSecretLayer4, SecretLayer4) {
     bool init() {
         if (!SecretLayer4::init()) return false;
 
-        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.875f);
+        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.75f);
 
         return true;
     };
@@ -43,7 +43,7 @@ class $modify(CbHookSecretLayer5, SecretLayer5) {
     bool init() {
         if (!SecretLayer5::init()) return false;
 
-        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.875f);
+        if (auto menu = typeinfo_cast<ClipboardMenu*>(getChildByIDRecursive("menu"_spr))) menu->setButtonScale(menu->getButtonScale() * 0.75f);
 
         return true;
     };