+++ changelog.md
@@ -1,5 +1,7 @@
 # v1.0.6 
 - Improve clipboard UI positioning with [FixThoseDangInputNodes!](https://geode-sdk.org/mods/cheeseworks.fixinputnodesizes)
+- Set default value for setting *`Button Opacity`* to `175`
+- Fixed some UI ordering issues
 
 # v1.0.5 
 - Ported to Geometry Dash version `2.2081`